- Added tem properties to entity objects until serialization is fixed

+++ wre.hubspot.apiclient/CRM/LineItems/IHubspotLineItem.cs
@@ -8,5 +8,5 @@ public interface IHubspotLineItem : IHubspotEntity
     decimal Discount { get; set; }
     decimal Quantity { get; set; }
     string? Name { get; set; }
-    int? ProductId { get; set; }
+    long? ProductId { get; set; }
 }
